Add new fontBoundingBoxAscent/Descent test
Add new fontBoundingBoxAscent/Descent test with Ahem font. Firefox and
Chrome follow different table to calculate the fontBoundingboxAscent and
descent. Ahem is one the font that has the same information across
the tables. In this cl, I add a new test using Ahem font, so that the
test can work for all browsers.

Bug:1129621

Change-Id: I8ef0d7c5788e79f215d763e8494c3db8941b95d8