[mathml] Make math display values on pseudo elements compute to flow
This is a follow-up of [1] which didn't handle the case of pseudo
elements with a math display value. A reftest is added to verify the
expected behavior with pseudo elements, but will fail until the
parsing of math display values is adjusted to match what the CSSWG
agreed. However, this tets checks the fix for an assertion failure
(http://crbug.com/1130127).

Bug: 6606, 1127222, 1130127

[1] https://chromium-review.googlesource.com/c/chromium/src/+/2404789

Change-Id: I4f7c13db003f776045a934eae9be5839b3c0343f